fix for secondary issue in #9
Added conditional bindings to `core-icon-button` for `menuButton` to position it to the right of content when `rightDrawer` is enabled.  This was to address a request from Polymer/core-scaffold/#9.

I didn't want to be too heavy-handed and force the button to be pushed to the right by adding a `<div flex></div>` before the `core-icon-button`, because it would prevent content from filling the entire area when using core-scaffold (since using flex in the content div would only split the width with the empty flex div inside core-scaffold).  This way is more flexible (pardon the pun) without having to modify core-scaffold.  Example here: http://jsbin.com/saweqesozuki/1/edit

Also, added some missing documentation and new documentation for this function.